requirements.txt
acres==0.5.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==24.10.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distlib==0.4.0
ecdsa==0.19.1
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.19.1
fitz==0.0.1.dev2
flake8==6.1.0
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
identify==2.6.13
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
looseversion==1.3.0
lxml==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
nodeenv==1.9.1
numpy==2.0.2
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.3.8
pre-commit==3.8.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
pyasn1==0.4.8
pycodestyle==2.11.1
pydantic==2.11.3
pydantic-settings==2.10.1
pydantic_core==2.33.1
pydot==4.0.1
pyflakes==3.1.0
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.5
rsa==4.9.1
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.2
virtualenv==20.34.0
zipp==3.23.0
